Add water script
This script needs to be put in something with a BoxCollider. It will
check if Medrash is inside it, and will limit his movement (changing
isOnWater to true, which changes canRun to false).

The modification inside the MainCharacterController will allow the
change of sound and inclusion of water ripples when the time comes.